update scripts to use newer versions of tools
the cli interface for refmt has changed, and now requires double-dashes (--) in
front of print and parse.

the cli interface for bs-platform no longer exposes a binary named bsc and
instead exposes a binary named bsb. bsb also doesn't need the same flags and
syntax that was in the previous command, and instead functions with `-make-world`