Response.send and some project updates (#5)
* update scripts to use newer versions of tools

the cli interface for refmt has changed, and now requires double-dashes (--) in
front of print and parse.

the cli interface for bs-platform no longer exposes a binary named bsc and
instead exposes a binary named bsb. bsb also doesn't need the same flags and
syntax that was in the previous command, and instead functions with `-make-world`

* don't track the generated .merlin file

the generated .merlin file contains user specific information. since this file
is generated with the paths present on the box it is generated, it breaks the
development experience for any other uses. instead, rely on bsb to generate this
file on any box it operates on.

* remove unused webpack devdep

* add binding to Response.send

* rewrite Response.send with externals overloading

* update Response.send types to restrict input to JSON rather than all 'a

* rename sendObject to sendJson, deprecate json

* Fix Typo