Set "files" property in package.json
I have an app that depends on bs-express. If I `npm install` the app, I'm unable to run it because `lib/js/src/Express.js` is missing from the published package. This PR should make it so that the compiled JS is included in the published package. As a side effect, extra files that don't need to be published (readme, tests, etc) will be excluded.