Migrate the example code to Rescript
@DerivedMate has ported bs-express to Rescript,
but the example is still in Reason.

I've ported it to Rescript syntax by using the
rescript playground to convert from reason to Rescript
and then hand editing the parts that "looked weird".

I have removed the original Reason file, but we could also
leave it in if you think there is benefit to having
two examples.